website/requirements.txt
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl
torchvision==0.2.2
Flask==1.0.3
Pillow==5.2.0
gunicorn==20.0.4
pickle
pandas